Improve setting static routes
The current implementation of setting static routes in vxlan-connector
only accepts device routes. If one wants to use a static route such as

    192.168.1.2 via 10.0.0.1 dev vx101

(next hop needs to go through the 'vx101' interface), one needs to use
the environment variable STATIC_ROUTES, set it like "192.168.1.2 via
10.0.0.1" and the _static_routes() function should take care of the
rest. Except that it does not:

    ip route add 192.168.1.2 via 10.0.0.1 dev vx101
    Error: any valid prefix is expected rather than "192.168.1.2 via 10.0.0.1".

The reason for this is that IFS is still set to "," when the variable
expansion is happening in the "ip route add" call.